Added README & Menu changes
Added guides to README.

Added buttons to the scenes in the Unity project.
Allowed the buttons to switch between different scenes!

PLEASE DO NOT CHANGE THIS YET - IT LOOKS TERRIBLE, BUT HAS BEEN DONE SO THAT I CAN BUILD PORTIONS OF THE GAME ACCESSED THROUGH THESE MENUS - THE ENTIRE BUTTON SYSTEM WILL BE REPLACED WITH SOMETHING BETTER LATER. Thanks :)