Minor QoL improvements
I fixed the jittering when colliding with the counter tops - also tweaked how to transition between locations. I've also significantly improved the performance of the game on PC (idk about android/ios yet) because Adaptive Perfomance was being troublesome. Currently it appears that the restaurant is getting a really low frame rate on build - I'm aware but unsure as how to currently fix this (TLDR: nothing major)