Split LevelOne/Two into Kitchen/Restaurant scenes
OK, Kitchen and Restaurant are split into seperate scenes, data preserved on scene swap. Need to be careful about objects persisting out of game, but to resolve l8r.There is an issue in the editor where the player sprite appears to warp to strange places on scene change, this doesnt appear in the build
additionally, the joystick doesnt continue with the current input on scene change. Not sure how to fix this atm - you just need to remove and re-click the widget.
Also added logo to start 🙂