Introducing customers
Customers now spawn in when the level starts, and on a timer (to be refined).

Customers stop spawning when there are 6 in the queue, as this is currently how many can fit on the screen.

Customers automatically move through the queue as it changes.

The first customer in the queue is identified and ready for interactions (tmrw)

Added level manager.

Level manager and customers aren't destroyed on scene change, but are hidden to their respective scenes.

TO-DO:
Customers generate meals
Pop-up to serve customers
Store meals
Link appliances in from there!